refactor: :recycle: Refactored the main page for a better folder structure
